add to cart

+++ be/src/main/java/matcha/project/be/dto/ItemResponseDto.java
@@ -4,11 +4,18 @@ import lombok.Data;
 
 @Data
 public class ItemResponseDto {
+
     private Long id;
+
     private String name;
+
     private String brand;
-    private Double price;
+
+    private Integer price;
+
     private String imageUrl;
+
     private Long amountSold;
+
     private Boolean isRecommended;
 }
